requirements.txt
pip==25.0.1
aiohttp
python-dotenv
pytest
requests